+++ CHANGES.md
@@ -5,6 +5,7 @@
 - validate & decode oidc token instead of blindly trusting it by querying userinfo endpoint
 - remove UserInfo class as a decoded token have sufficient identity information
 - added relationships object in model responses
+- added relationship traversal in client
 
 ## 0.1.2b3 (2023-08-24)